summarizer/requirements.txt
slack-sdk>=3.23.0
PyGithub>=2.1.1
anthropic>=0.18.0
google-generativeai>=0.3.0
